requirements.txt
streamlit==1.12.2
Pillow==9.2.0
cairosvg==2.5.2